minor fix: 64 bits build, missing std:: in header file

+++ libdb/db_debug.c
@@ -76,7 +76,7 @@ static int check_redundant_key(samples_odb_t const * hash, odb_key_t max)
 	for (pos = 1 ; pos < hash->descr->current_size ; ++pos) {
 		if (bitmap[hash->node_base[pos].key]) {
 			printf("redundant key found %lld\n",
-			       hash->node_base[pos].key);
+			       (unsigned long long)hash->node_base[pos].key);
 			return 1;
 		}
 		bitmap[hash->node_base[pos].key] = 1;
